+++ src/Color.hs
@@ -14,7 +14,7 @@ cAdd (Color r1 g1 b1) (Color r2 g2 b2) = Color (r1 + r2) (g1 + g2) (b1 + b2)
 cSub :: Color -> Color -> Color
 cSub (Color r1 g1 b1) (Color r2 g2 b2) = Color (r1 - r2) (g1 - g2) (b1 - b2)
 
-cMult :: Float -> Color -> Color
+cMult :: Double -> Color -> Color
 cMult c (Color r g b) = Color (c * r) (c * g) (c * b)
 
 hadamard :: Color -> Color -> Color
