+++ src/Ray.hs
@@ -4,7 +4,7 @@ import Types ( Point, Transformation, Ray(..) )
 import Transformation (mpMult, mvMult)
 import VecPoint (vMult, vpAdd)
 
-position :: Ray -> Float -> Point
+position :: Ray -> Double -> Point
 position (Ray origin direction) t = origin `vpAdd` (t `vMult` direction)
 
 transform :: Ray -> Transformation -> Ray
