+++ src/VecPoint.hs
@@ -2,7 +2,7 @@ module VecPoint where
 
 import Types (Point (..), Vec (..))
 
-epsilon :: Float
+epsilon :: Double
 epsilon = 0.00001
 
 {- Point functions -}
@@ -12,10 +12,10 @@ pAdd (Point x1 y1 z1) (Point x2 y2 z2) = Vec (x1 + x2) (y1 + y2) (z1 + z2)
 pSub :: Point -> Point -> Vec
 pSub (Point x1 y1 z1) (Point x2 y2 z2) = Vec (x1 - x2) (y1 - y2) (z1 - z2)
 
-pMult :: Float -> Point -> Point
+pMult :: Double -> Point -> Point
 pMult c (Point x y z) = Point (c * x) (c * y) (c * z)
 
-pDiv :: Point -> Float -> Point
+pDiv :: Point -> Double -> Point
 pDiv (Point x y z) c = Point (x / c) (y / c) (z / c)
 
 {- Basic vector functions -}
@@ -28,10 +28,10 @@ vSub (Vec x1 y1 z1) (Vec x2 y2 z2) = Vec (x1 - x2) (y1 - y2) (z1 - z2)
 vNeg :: Vec -> Vec
 vNeg (Vec x y z) = Vec (- x) (- y) (- z)
 
-vMult :: Float -> Vec -> Vec
+vMult :: Double -> Vec -> Vec
 vMult c (Vec x y z) = Vec (c * x) (c * y) (c * z)
 
-vDiv :: Vec -> Float -> Vec
+vDiv :: Vec -> Double -> Vec
 vDiv (Vec x y z) c = Vec (x / c) (y / c) (z / c)
 
 {- Point-vector functions -}
@@ -42,7 +42,7 @@ vpSub :: Point -> Vec -> Point
 vpSub (Point x1 y1 z1) (Vec x2 y2 z2) = Point (x1 - x2) (y1 - y2) (z1 - z2)
 
 {- Other vector functions -}
-magnitude :: Vec -> Float
+magnitude :: Vec -> Double
 magnitude (Vec x y z) = sqrt ((x ^ 2) + (y ^ 2) + (z ^ 2))
 
 normalize :: Vec -> Vec
@@ -50,7 +50,7 @@ normalize v@(Vec x y z) = Vec (x / mag) (y / mag) (z / mag)
   where
     mag = magnitude v
 
-dot :: Vec -> Vec -> Float
+dot :: Vec -> Vec -> Double
 dot (Vec x1 y1 z1) (Vec x2 y2 z2) = (x1 * x2) + (y1 * y2) + (z1 * z2)
 
 cross :: Vec -> Vec -> Vec
